NR#01-Client: Refactoring cookie flash message to listen to incoming notification that the notification is show
To ensure once user clicks the notification close button or the notification is auto removed from the UI, when the notification originated from the server, a notification is sent to remove that message from the flash messages stored in the cookie (FlashMessages). This was the first implementation. What remains is to implement filtering the notification and actual implementation of sending messages from the client.

Other things implemented in this commit is renaming of the word Bug or bug to Bag or bag. Was an honest grammatical typo.